Improve popup positioning and value calculation. Add popup properties property to line data class to be able to set different popup properties in each line. Code improvement.

+++ compose-charts/src/commonMain/kotlin/ir/ehsannarmani/compose_charts/ColumnChart.kt
@@ -248,11 +248,11 @@ fun ColumnChart(
 
                 val barsAreaWidth = size.width - (indicatorAreaWidth)
                 val zeroY = size.height - calculateOffset(
-                    maxValue = maxValue.toFloat(),
-                    minValue = minValue.toFloat(),
+                    maxValue = maxValue,
+                    minValue = minValue,
                     total = size.height,
                     value = 0.0f
-                )
+                ).toFloat()
 
 
                 if (indicatorProperties.enabled) {
